* Delphi compatibility.
git-svn-id: https://mseide-msegui.svn.sourceforge.net/svnroot/mseide-msegui/trunk@3843 56ecf5fe-6917-0410-8ac6-ca8e178bccbf

+++ lib/common/kernel/msedatalist.pas
@@ -5262,7 +5262,7 @@ begin
   items[index]:= emptyreal;
  end
  else begin
-  if trystrtofloat(avalue,rea1) then begin
+  if trystrtofloat(avalue,double(rea1)) then begin
    items[index]:= rea1;
   end;
  end;
